AgenticWorkers/VectorDBWorker/requirements.txt
azure-storage-queue==12.9.0
azure-storage-blob==12.19.0
pinecone-client==3.0.0
sentence-transformers==2.3.1
python-dotenv==1.0.0
